+++ src/GOclasses/basic/MigrationSelectionPolicy.cpp
@@ -31,7 +31,7 @@ int MigrationSelectionPolicy::getNumberOfIndividualsToMigrate(const Population& 
 		}
 		return (int)(migrationRateFrac * (double)population.size());
 	} else {
-		if(migrationRateAbs > population.size()) {
+                if((size_t)migrationRateAbs > population.size()) {
 			pagmo_throw(assertion_error, "Absolute migration rate exceeds population size!");
 		}
 		return migrationRateAbs;
